==> R/2_PCS_data_formatting.Rmd <==
---
title: 'format PCS data'
author: 'Alex Huddell'
date: '10/7/2021'
output:
  word_document: default
  html_document: default
  pdf_document: default
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#Load libraries and data

```{r load libs}
library(tidyverse)
library(lubridate)
library(here)

dat<-read_csv(file=unzip(zipfile=here("data","PCS_LIS_complete_data.zip")))

#read in CT data from N exchange program
CT<-read_csv(file=here("data","CT_NX_data.csv"))
CT<-distinct(CT) #remove a few duplicate rows
CT_permits<-unique(CT$permit)


#join by location and SIC code to filter out SIC codes other than sewarage systems, fish hatcheries, and paper mills 
outfall_locations <-
  read_csv(file = here("data", "npdes_outfalls_layer.csv"))
names(outfall_locations)
names(dat)
outfall_locations<-rename(outfall_locations,"Permit Num (NPID)"=EXTERNAL_PERMIT_NMBR)

join<-left_join(dat,outfall_locations)

# SIC_desc<-join %>% group_by(SIC_CODES, SIC_DESCRIPTIONS) %>%
#   summarize(first(SIC_CODES))
# write.csv(SIC_desc, "SIC codes PCS.csv", row.names = F)
          
```


#Create key for each outfall/date and check how many rows per key.

```{r create key, echo=FALSE}
dat$date<-ymd(dat$`Monitoring Dt (MVDT)`)

dat <- dat %>%
  mutate(days_per_month = days_in_month(date)) %>%
  mutate(key = paste0(`Permit Num (NPID)`,`Discharge Num (DSCH)`,date))

nrow(dat) #there are 196,586 rows
dat %>% count(key) #there are 95,366 unique keys, so some keys have duplicate rows

#plot the frequency of different parameters
dat$param<-as.factor(dat$'Parameter Cd (PRAM)')
levels(dat$param)

levels<-dat %>%
  group_by(param) %>%
  tally()

ggplot(levels, aes(param,n))+
  geom_col() +
  ylab('count of observations')+
  theme(axis.text.x=element_text(angle = 30, hjust=1))



levels<-dat %>%
  group_by(param) %>%
  tally()

ggplot(levels, aes(param,n))+
  geom_col() +
  ylab('count of observations')+
  theme(axis.text.x=element_text(angle = 30, hjust=1))
```

```{r rename, organize, and filter}

#renaming columns  
dat<-dat%>%
  rename(facility=`Facility Name Short (FNMS)`,
         permit=`Permit Num (NPID)`,
         Outfall=`Discharge Num (DSCH)`,
         quant_avg=`Meas Quant Avge (MQAV)`,
         conc_avg=`Meas Conc Avge (MCAV)`) %>%
# reformatting date and add a key and number of days in month column, then filter out some problematic units
  mutate(key = paste0(permit, '_', Outfall, '_', date)) %>%
  mutate(days_per_month = as.numeric(days_in_month(date))) %>%
  mutate(permit_outfall = as.factor(paste0(permit, "_", Outfall))) %>%
  mutate(permit_outfall_designator=as.factor(paste0(permit,"_",Outfall,"_",`Discharge Designator (DRID)`))) %>%
  mutate(state=substr(permit,1,2)) #add column with state abbreviation 


##read outfall locations data, add key, and join to dat so we can filter by SIC code
outfall_locations <-
  read_csv(file = here("data", "npdes_outfalls_layer.csv"))

outfall_locations$permit<-outfall_locations$EXTERNAL_PERMIT_NMBR

join<-left_join(dat,outfall_locations, by='permit')

SIC_permit<-join %>%
  group_by(permit) %>%
  summarize(SIC=first(SIC_DESCRIPTIONS),
            SIC_CODES=first(SIC_CODES))

SIC_permit_count<-SIC_permit %>%
  group_by(SIC,SIC_CODES) %>%
  tally()%>%
  arrange(desc(n))

SIC_permit_count %>%
  filter(n>5)%>%
  ggplot(aes(SIC ,n))+
  geom_col() +
  ylab("count of observations")+
  theme(axis.text.x=element_text(angle = 60, hjust=1))


#list of SIC codes to keep?
#	Sewerage Systems--4952
# Colleges And Universities; Elementary And Secondary Schools--8221
#Services; Sewerage Systems--8999 

SIC_CODES_keep<-c('4952','8221','8999', '4952 8221', '9512 4952',
                  '8999 4952')

permits_to_keep<-join %>%
  filter(SIC_CODES %in% SIC_CODES_keep) %>%
  group_by(permit)%>%
  summarize(permit=first(permit),
            facility=first(facility))

#filtering out unwanted observations
dat<- dat %>% 
  filter(permit %in% (permits_to_keep$permit)) %>%  
  filter(!permit %in% c('CT0000434','CT0026476','NY0065340','NY0066028')) %>% #removing some industrial permits that seem unrelated to sewerage systems, but some small private WWTPs are still included (like hospitals)
  #removing unwanted discharge codes and monitoring locations
  filter(`Monitoring Location (MLOC)` %in% c("1", "2", "EA", "EG", "Y")) %>%
  #removing one unclear unit that only relates to a few observations of flow maxima
  filter(!`Quant Unit Cd Desc (LQUCD)` =='(3R) - Million Gallons')%>%
  #omitting this parameter because we don't know what it means, but most
  filter(!param=='50047 - Flow, Maximum During 24 Hr Period') %>%
  filter(!param=='00059 - Flow Rate, Instantaneous') %>%
  filter(!param=='51061 - Flow (dry weather)') %>% #there are only a few observations of this and none had data we could use
  mutate(permit_outfall_designator_key=paste0(permit_outfall_designator,key)) %>%
  filter(!permit_outfall_designator_key=='NY0026841_1_MNY0026841_1_2004-05-31' ) %>% #the M designator causes a duplicate this month
  filter(!permit_outfall_designator_key=='NY0026841_1_MNY0026841_1_2004-06-30' ) %>% #the M designator causes a duplicates this month
  mutate(key_quant_avg=paste0(key,quant_avg)) %>%
  filter(!key_quant_avg=='CT0100081_1_1989-12-311.83') %>% #filtering out one duplicate row
  select(-key_quant_avg, -permit_outfall_designator_key) %>%
  filter(!permit %in% CT_permits) %>%#this line removes the CT waste water treatment plants because we will use CTDEEP data for those
  
  # #add in all of the specific cases we need to omit to eliminate duplicates
  mutate(permit_outfall_ML=paste0(permit_outfall, `Monitoring Location (MLOC)`)) %>%
  filter(!permit_outfall_ML == 'NY0026697_1Y')  %>%  #there are duplicates for ML 1 and Y, choosing 1
  filter(!permit_outfall_ML == 'NY0026701_1Y')  %>%  #there are duplicates for ML 1 and Y, choosing 1
  filter(!permit_outfall_ML == 'NY0026719_1Y')  %>%  #there are duplicates for ML 1 and Y, choosing 1
  filter(!permit_outfall_ML == 'NY0026786_1Y')  %>%  #there are duplicates for ML 1 and Y, choosing 1

  
    select(-permit_outfall_ML) 

  # filter(!permit_outfall_designator == 'CT0003824_001_A')  %>%  #this one I'm not sure, but for ECHO data we chose C; here it seems the measurements vary across designators more from month to month
  # filter(!permit_outfall_designator == 'CT0003824_001_B')  %>%  #this one I'm not sure, but for ECHO data we chose C; here it seems the measurements vary across designators more from month to month
  # filter(!permit_outfall_designator == 'CT0003824_001_D')  %>%  #this one I'm not sure, but for ECHO data we chose C; here it seems the measurements vary across designators more from month to month
  # filter(!permit_outfall_designator == 'CT0003824_001_E')  %>%  #this one I'm not sure, but for ECHO data we chose C; here it seems the measurements vary across designators more from month to month
  # filter(!permit_outfall_designator == 'CT0003824_002_A')  %>%  #this one I'm not sure, but for ECHO data we chose C; here it seems the measurements vary across designators more from month to month
  # filter(!permit_outfall_designator == 'CT0003824_002_C')  #this one I'm not sure, but for ECHO data we chose C; here it seems the measurements vary across designators more from month to month

dat$`Quant Unit Cd Desc (LQUCD)` <- droplevels(dat)$`Quant Unit Cd Desc (LQUCD)` # dropping unused levels from dat$`Quant Unit Cd Desc (LQUCD)`
dat$`Conc Unit Cd Desc (LCUCD)` <- droplevels(dat)$`Conc Unit Cd Desc (LCUCD)` # dropping unused levels from dat$`Quant Unit Cd Desc (LQUCD)`
dat$param<- droplevels(dat)$param # dropping unused levels from dat$param

#make values numeric
dat$quant_avg<-as.numeric(dat$quant_avg)
dat$conc_avg<-as.numeric(dat$conc_avg)


dim(dat)

length(unique(as.character(dat$permit_outfall))) #189 permit/outfalls
length(levels(as.factor(dat$key))) #22,970 keys
```


In the next chunk, we reduce the number of units to three--one for flow, one for 
concentration, and another for loading rate. 

```{r standardize units}
#standardize units to kg per month, million L per month, or milligram per liter
dat <- dat %>% mutate(
  kg_per_month =
    case_when(
      `Quant Unit Cd Desc (LQUCD)` == '(26) - Pounds per Day' ~ quant_avg / 2.205 * days_per_month,
      `Quant Unit Cd Desc (LQUCD)` == '(01) - Kilograms per Day' ~ quant_avg  * days_per_month # converts per day to per month
             )
)

#standardize units to millions of L per month
dat <- dat %>% mutate(
  million_L_per_month =
    case_when(
      `Quant Unit Cd Desc (LQUCD)` == '(07) - Gallons per Day' ~ quant_avg / 10 ^ 6 * 3.785 * days_per_month,
      #converts from Gallons per day to millions L per month
      `Quant Unit Cd Desc (LQUCD)` == '(08) - Cubic Feet per Second' ~ quant_avg * 28.317 / 10 ^ 6 * 60 * 60 * 24 * days_per_month,
      # converts cubic feet per second to million L per month
      
      `Quant Unit Cd Desc (LQUCD)` == '(8D) - Gallons per Month' ~ quant_avg / 10 ^ 6 * 3.785,
      # converts gallons to millions of L
      `Quant Unit Cd Desc (LQUCD)` == '(03) - Million Gallons per Day' ~ quant_avg  * 3.785 * days_per_month,
      # converts gallons to L and day to month
      `Quant Unit Cd Desc (LQUCD)` == '(78) - Gallons per Minute' ~ quant_avg  / 10 ^ 6 * 3.785 * 60 * 24 * days_per_month,
      # converts gallons to millions of L and minute to month
    )
)

#converting all N parameters to mass terms of nitrogen and not others (i.e. ammonia, nitrite, etc.)


dat$conc_avg <-  ifelse(
  dat$param == '00608 - Nitrogen, ammonia dissolved',
  dat$conc_avg * 14.0067 / 17.031,
  dat$conc_avg
)

dat$conc_avg <- ifelse(
  dat$param == '34726 - Nitrogen, ammonia, total (as NH3)',
  dat$conc_avg * 14.0067 / 17.031,
  dat$conc_avg
)

dat$conc_avg <- ifelse(
  dat$param ==  '71850 - Nitrogen, Nitrate Total (As No3)' ,
  dat$conc_avg * 14.0067 / 62.0049,
  dat$conc_avg
)

   
#standardize concentrations to milligrams N per liter  
dat <- dat %>% mutate(
  mg_N_per_L=
    case_when(
   `Conc Unit Cd Desc (LCUCD)` == 'Milligrams Per Liter' ~ conc_avg,
   `Conc Unit Cd Desc (LCUCD)` == 'Micrograms Per Liter' ~ conc_avg / 1000, #convert from micrograms to milligram
    )
  )



```
Next, we reduce the number of nitrogen to the following: total nitrogen 'TN',
organic nitrogen 'ON', ammonia 'NH3', nitrite 'NO2', nitrate 'NO3', and inorganic N 'IN'.
We reduce the flow parameters to the following: flow rate 'flow_rate', 
maximum flow rate 'flow_max_24h', total flow 'flow_total', and dry flow 'flow_dry'.

```{r simplify parameters}
#cleaning up duplicate nitrogen and flow parameters 
dat$param<-as.factor(dat$param)

         
#dat<-dat %>% mutate(param_simple=recode(dat$param,
#cleaning up duplicate nitrogen and flow parameters 
dat$param_simple <- recode(dat$param,

    # #Nitrogen parameters
    '00600 - Nitrogen, Total (As N)' = 'TN',
    '00605 - Nitrogen, Organic Total (As N)' = 'ON',
    '00608 - Nitrogen, Ammonia Dissolved' = 'NH3',
    '00610 - Nitrogen, Ammonia Total (As N)' = 'NH3',
    '00615 - Nitrogen, Nitrite Total (As N)' = 'NO2',
    '00620 - Nitrogen, Nitrate Total (As N)' = 'NO3',
    '00625 - Nitrogen, Kjeldahl, Total (As N)'='TKN',
    '00630 - Nitrite Plus Nitrate Total 1 Det. (As N)' = 'NO2NO3',
    '34726 - Nitrogen, Ammonia, Total (As Nh3)' = 'NH3',
    '49579 - Nitrogen, Total Kjeldahl' = 'TKN',
    '51087 - Nitrogen, Kjeldahl, Total (Tkn) (Water)' = 'TKN',
    '71850 - Nitrogen, Nitrate Total (As No3)' ='NO3',
    #flow parameters
    '50050 - Flow, In Conduit Or Thru Treatment Plant' ='flow_rate',
    '51061 - Flow (Dry Weather)' = 'flow_rate',
    '00056 - Flow Rate' = 'flow_rate',
    '74076 - Flow' = 'flow_rate',
    '00058 - Flow Rate' = 'flow_rate',
    '82220 - Flow, Total' = 'flow_total',

    
  )

```

Let's separate the keys for which kg N/month loads already exist from those which don't.

```{r separate observations reporting loads}
#review duplicates
dat %>%
  filter(!is.na(kg_per_month)) %>%
  filter(param_simple=='TN') %>%
  group_by(key)%>%
  tally() %>%
  arrange(desc(n))
#there are no duplicates

#subset of data that report loads
N_load<- dat %>%
  group_by(key) %>%
  filter(!is.na(kg_per_month)) %>%
    #removing unwanted columns
  select(
    !c('Monitoring Dt (MVDT)',
        'Discharge Designator (DRID)', 
       'Parameter Cd (PRAM)',
       'Monitoring Location (MLOC)',
       '#MULTIVALUE',   
       'Meas Quant Maxm (MQMX)', 
       'Quant Unit Cd (LQUC)',
       'Quant Unit Cd Desc (LQUCD)',
       'Meas Conc Minm (MCMN)',
       'Conc Unit Cd (LCUC)',
       'Conc Unit Cd Desc (LCUCD)',
       'param')
  ) %>%
  select(!starts_with('Meas Conc')) %>%
  distinct() #removing any duplicate rows


#check if there are duplicates per key and one parameter
N_load %>%
  group_by(key)%>%
  filter(param_simple=='TN') %>%
  tally() %>%
  arrange(desc(n))
# there are no duplicates

unique(N_load$param_simple)

#pivoting the TN and NH3 observations into wide format and removing flow and concentration columns
N_load_a<-N_load %>%
   pivot_wider(names_from = param_simple,
              values_from = kg_per_month) %>%
   select(-million_L_per_month,- mg_N_per_L, -permit_outfall_designator) %>%
  rename(kg_N_TN_per_month=TN,
         kg_N_NH3_per_month=NH3,
        kg_N_NO3_per_month=NO3, 
        kg_N_NO2_per_month=NO2, 
        kg_N_NO2NO3_per_month=NO2NO3,
        kg_N_TKN_per_month=TKN)

names(N_load_a)
length(unique(as.character(N_load_a$permit_outfall))) #36 outfalls included here 
```

Here we pivot the data wider calculate loads based on the concentrations and flow data, and remove the concentration and flow columns after we use them.


```{r calculating loads for other observations}
#function to id columns with all NAs
not_all_na <- function(x) any(!is.na(x))

#review duplicates
keys_with_N<- dat %>%
  group_by(key) %>%
  filter(!any(!is.na(kg_per_month)))%>% #this removes any keys which do report in N loads
  filter(any(as.character(param) %in% c(
      '00625 - Nitrogen, Kjeldahl, Total (As N)',
      '00605 - Nitrogen, Organic Total (As N)',
      '00610 - Nitrogen, Ammonia Total (As N)',
      '00615 - Nitrogen, Nitrite Total (As N)',
      '00620 - Nitrogen, Nitrate Total (As N)',
      '71850 - Nitrogen, Nitrate Total (As No3)',
      '00600 - Nitrogen, Total (As N)',
      '34726 - Nitrogen, Ammonia, Total (As Nh3)',
      '00630 - Nitrite Plus Nitrate Total 1 Det. (As N)',
      '00608 - Nitrogen, Ammonia Dissolved',
      '51087 - Nitrogen, Kjeldahl, Total (Tkn) (Water)',
      '49579 - Nitrogen, Total Kjeldahl'
    ))) #this removes any keys which do report any other N parameters

keys_with_N


dup<-keys_with_N %>%
  group_by(key) %>%
  drop_na(mg_N_per_L)%>%
  group_by(key, param_simple)%>%
  tally() %>%
  filter(n>1)%>%
    arrange(desc(n))
dup #there aren't any duplicates after excluding NAs



#separate out concentration data so that we can drop keys empty concentrations
N_params<- keys_with_N %>%
group_by(key) %>%
  filter(!any(!is.na(kg_per_month)))%>% #this removes any keys which do report in N loads
  filter(!param == '50050 - Flow, In Conduit Or Thru Treatment Plant') %>%
  filter(!param == '00056 - Flow Rate') %>%
  filter(!param == '74076 - Flow') %>%
  filter(!param == '00058 - Flow rate') %>%
  filter(!param == '51061 - Flow (dry weather)') %>%
  drop_na(mg_N_per_L)%>%
distinct()

flow_all<- keys_with_N %>%
group_by(key) %>%
  filter(!any(!is.na(kg_per_month)))%>% #this removes any keys which do report in N loads
  filter(param %in% c('50050 - Flow, In Conduit Or Thru Treatment Plant', 
                      '00056 - Flow Rate', '74076 - Flow',
                      '00058 - Flow rate','51061 - Flow (dry weather)') 
         ) %>%
  distinct()

#bind all observations back together
 flow_cleaned<-rbind(flow_all,N_params)
 
 flow_cleaned <- flow_cleaned %>%
  #removing unwanted columns
  select(
    !c('Monitoring Dt (MVDT)',
        'Discharge Designator (DRID)',
       'Parameter Cd (PRAM)',
       'param',
       'Monitoring Location (MLOC)',
       '#MULTIVALUE',
       'Meas Quant Maxm (MQMX)',
       'Quant Unit Cd (LQUC)',
       'Quant Unit Cd Desc (LQUCD)',
       'Meas Conc Minm (MCMN)',
       # 'Meas Conc Maxm (MCMX)_1',
       # 'Meas Conc Maxm (MCMX)',
       'Conc Unit Cd (LCUC)',
       'Conc Unit Cd Desc (LCUCD)',
       'quant_avg',
       'permit_outfall_designator',
       'conc_avg'
       )
  ) %>%
  select(!starts_with('Meas Conc Maxm (MCMX')) %>%

  distinct() 
 
 dup<-flow_cleaned %>%
  group_by(key, param_simple)%>%
  tally() %>%
  filter(n>1)%>%
    arrange(desc(n))
dup #there aren't any duplicates after excluding NAs

```

Calculate loads for keys which do not report loads.

```{r}
#subset of data that DO NOT report loads, then
#pivot wider and then remove columns with only NAs (since some parameter/unit combinations that don't make sense)

flow_N_concentration <- flow_cleaned %>%
  group_by(key) %>%
  pivot_wider(names_from = param_simple,
              values_from = c(mg_N_per_L, million_L_per_month)) %>%
  select_if(not_all_na)

#review duplicates
dups3<-flow_N_concentration %>%
  group_by(key)%>%
  tally() %>%
  arrange(desc(n))
dups3 #there are no more duplicates


#calculating N loads
N_load_b <- flow_N_concentration %>%
  mutate(kg_N_TN_per_month = mg_N_per_L_TN * million_L_per_month_flow_rate) %>%
  mutate(kg_N_ON_per_month = mg_N_per_L_ON * million_L_per_month_flow_rate) %>%
  mutate(kg_N_NH3_per_month = mg_N_per_L_NH3 * million_L_per_month_flow_rate) %>%
  mutate(kg_N_NO2_per_month = mg_N_per_L_NO2 * million_L_per_month_flow_rate) %>%
  mutate(kg_N_NO3_per_month = mg_N_per_L_NO3 * million_L_per_month_flow_rate) %>%
   mutate(kg_N_TKN_per_month = mg_N_per_L_TKN * million_L_per_month_flow_rate) %>%
  mutate(kg_N_NO3NO2_per_month = mg_N_per_L_NO2NO3 * million_L_per_month_flow_rate) %>%
  select(
    -mg_N_per_L_TN,
    -mg_N_per_L_NH3,
    -mg_N_per_L_NO2,
    -mg_N_per_L_NO3,
    -mg_N_per_L_NO2NO3,
    -mg_N_per_L_TKN,
    -mg_N_per_L_ON,
    -million_L_per_month_flow_rate,
  )

#combine N columns in N_load_b to standardize and join dataframes
names(N_load_b)


#check if any rows contain both NO3 or NO2 and NO3NO2 before we combine them
filter(N_load_b, !is.na(kg_N_NO3_per_month) & !is.na(kg_N_NO3NO2_per_month))
filter(N_load_b, !is.na(kg_N_NO2_per_month) & !is.na(kg_N_NO3NO2_per_month))
#there are none, so we can add them without double counting

#sum up other forms of N apart from TN
N_load_b <- N_load_b %>%
  mutate(kg_N_sum_per_month =
           case_when(
             all(
               is.na(kg_N_NH3_per_month),
               is.na(kg_N_NO3_per_month),
               is.na(kg_N_NO2_per_month),
               is.na(kg_N_NO3NO2_per_month)
               ) ~ NA_real_,
             is.na(kg_N_TKN_per_month) ~ sum(
               kg_N_NH3_per_month,
               kg_N_NO3_per_month,
               kg_N_NO2_per_month,
               kg_N_NO3NO2_per_month,
               na.rm = TRUE
             ),
             !is.na(kg_N_TKN_per_month) ~ sum(
               kg_N_TKN_per_month,
               kg_N_NO3_per_month,
               kg_N_NO2_per_month,
               kg_N_NO3NO2_per_month,
               na.rm = TRUE
             )
           )) 

#remove intermediate columns we no longer need
N_load_b <- N_load_b %>%
  select(
    -kg_N_NH3_per_month,-kg_N_NO3_per_month,
    -kg_N_NO2_per_month,
    -kg_N_ON_per_month,
    -kg_N_NO3NO2_per_month,
    -kg_N_ON_per_month,
    -kg_N_TKN_per_month
  )

#combine the different N columns; use total N when reported, or the sum of all other
#forms of N if TN is not reported
N_load_b$kg_N_TN_per_month <-
  ifelse(!is.na(N_load_b$kg_N_TN_per_month),
    N_load_b$kg_N_TN_per_month,
    N_load_b$kg_N_sum_per_month )

#remove N sum column
N_load_b<- N_load_b %>%
  select(-kg_N_sum_per_month)

names(N_load_b)
length(unique(as.character(N_load_b$permit_outfall))) #130 outfalls included here


#review duplicates
dups4<-N_load_b %>%
  group_by(key)%>%
  tally() %>%
  arrange(desc(n))
dups4 #there are no more duplicates


#for N_load_a
#combine the different N columns; use total N when reported, or the sum of all other
#forms of N if TN is not reported

#check if any rows contain both NO3 or NO2 and NO3NO2 before we combine them
filter(N_load_a, !is.na(kg_N_NO3_per_month) & !is.na(kg_N_NO2NO3_per_month))
#there are none, so we can add them without double counting
filter(N_load_a, !is.na(kg_N_NO2_per_month) & !is.na(kg_N_NO2NO3_per_month)) #14 rows have both, so we have to subtract NO2 from kg_N_NO3NO2_per_month for those rows

N_load_a$kg_N_sum_per_month <-
  ifelse(!is.na(N_load_a$kg_N_TKN_per_month),
    (N_load_a$kg_N_TKN_per_month + N_load_a$kg_N_NO3_per_month +        N_load_a$kg_N_NO2_per_month + N_load_a$kg_N_NO2NO3_per_month ),
    (N_load_a$kg_N_NH3_per_month + N_load_a$kg_N_NO3_per_month +   N_load_a$kg_N_NO2_per_month +  N_load_a$kg_N_NO2NO3_per_month )
  )

length(unique(as.character(N_load_a$permit_outfall))) #36 outfalls included here


#remove intermediate columns we no longer need
N_load_a <- N_load_a %>%
  select(
    -kg_N_NH3_per_month,
    -kg_N_NO3_per_month,
    -kg_N_NO2_per_month,
    -kg_N_NO2NO3_per_month,
    -kg_N_TKN_per_month
  )

#combine the different N columns; use total N when reported, or the sum of all other
#forms of N if TN is not reported

N_load_a$kg_N_TN_per_month <-  ifelse(!is.na(N_load_a$kg_N_TN_per_month),
    N_load_a$kg_N_TN_per_month,
    N_load_a$kg_N_sum_per_month )

#remove other unnecessary columns
N_load_a<- N_load_a %>%
  select(-kg_N_sum_per_month)
names(N_load_a)


#combine dataframes
N_load<-rbind(N_load_a,N_load_b)
N_load<-drop_na(N_load,kg_N_TN_per_month) #dropping NA values 

length(unique(as.character(N_load$permit_outfall))) #51 outfalls with/without locations
length(unique(as.character(N_load$key))) #3,891 keys

summary(N_load$kg_N_TN_per_month)

N_load<- N_load %>%
  mutate(month_year=format(as.Date(date), "%Y-%m")) %>%
  select(-date)


```


Loading in NPDES database with NPDES ID to match locations and other data.

```{r outfall location join}
outfall_locations <-
  read_csv(file = here("data", "npdes_outfalls_layer.csv"))

names(outfall_locations)

outfall_locations<-outfall_locations %>%
  select(EXTERNAL_PERMIT_NMBR,
         FACILITY_NAME,
         SIC_CODES,
         SIC_DESCRIPTIONS,
         PERM_FEATURE_NMBR,
         STATE_WATER_BODY_NAME,
         LATITUDE83,
         LONGITUDE83)


outfall_locations$permit_outfall <-
  paste0(outfall_locations$EXTERNAL_PERMIT_NMBR,
         "_",
         outfall_locations$PERM_FEATURE_NMBR)

#join locations based on permit number and outfall number
dim(N_load) #3891
length(unique(N_load$permit_outfall)) #51 outfalls with/without locations
length(unique(N_load$key)) #3891 keys

N_load_out <- left_join(N_load, outfall_locations, by = "permit_outfall")
dim(N_load_out) #3891

length(unique(as.character(N_load_out$permit_outfall))) #51 outfalls with locations
length(unique(as.character(N_load_out$key))) #3891 keys
#same data after join

#check for NA's in latitude
summary(N_load_out$LATITUDE83) #80 missing values
missing_latitude<-N_load_out %>% filter(is.na(LATITUDE83)) %>%
  group_by(permit_outfall)%>%
  summarize(LATITUDE83=first(LATITUDE83))
missing_latitude


#impute missing values based on other outfalls from same permit
N_load_out<-N_load_out %>% 
  mutate(impute_lat= case_when(
    permit_outfall=='MA0100889_FAC' ~ 42.190537,
    permit_outfall=='MA0101281_1' ~ 42.561923,
    permit_outfall== 'MA0101613_1'~	42.085814,	            
    permit_outfall== 'MA0101681_2'~	42.404149,	            
    permit_outfall== 'MA0101818_2'~	42.30248,	            
    permit_outfall== 'MA0110264_1'~	42.606111,
    permit_outfall=='NH0000744_12' ~ 44.323049,
    permit_outfall=='NH0000744_13' ~ 44.323049,
    permit_outfall=='NH0000744_14' ~ 44.323049,
    permit_outfall=='NH0000744_15' ~ 44.323049,
    permit_outfall=='NH0000744_16' ~ 44.323049,
    permit_outfall=='NH0000744_17' ~ 44.323049,
    permit_outfall=='NH0000744_18' ~ 44.323049,
    permit_outfall=='NH0000744_19' ~ 44.323049,
    permit_outfall=='NH0000744_20' ~ 44.323049,
    permit_outfall=='NH0000744_22' ~ 44.323049,
    permit_outfall=='NH0000744_23' ~ 44.323049,
    permit_outfall=='NH0000744_24' ~ 44.323049,
    permit_outfall== 'NY0021750_2' ~ 40.948583,	
    permit_outfall== 'NY0206644_2' ~ 40.924472,
    permit_outfall== 'NH0100307_FAC' ~ 43.374692)) %>%
   mutate(impute_lon= case_when(
    permit_outfall=='MA0100889_FAC' ~ -72.250278,
    permit_outfall=='MA0101281_1'~ -72.845629,
    permit_outfall== 'MA0101613_1' ~	-72.584513,
    permit_outfall== 'MA0101681_2' ~	-73.2412,	            
    permit_outfall== 'MA0101818_2' ~ -72.607846,
    permit_outfall== 'MA0110264_1' ~	-72.534861,
    permit_outfall=='NH0000744_12' ~ -71.560259,
    permit_outfall=='NH0000744_13' ~ -71.560259,
    permit_outfall=='NH0000744_14' ~ -71.560259,
    permit_outfall=='NH0000744_15' ~ -71.560259,
    permit_outfall=='NH0000744_16' ~ -71.560259,
    permit_outfall=='NH0000744_17' ~ -71.560259,
    permit_outfall=='NH0000744_18' ~ -71.560259,
    permit_outfall=='NH0000744_19' ~ -71.560259,
    permit_outfall=='NH0000744_20' ~ -71.560259,
    permit_outfall=='NH0000744_22' ~ -71.560259,
    permit_outfall=='NH0000744_23' ~ -71.560259,
    permit_outfall=='NH0000744_24' ~ -71.560259,
    permit_outfall== 'NY0021750_2' ~ -73.075639,
    permit_outfall== 'NY0206644_2' ~ -73.121139,
    permit_outfall== 'NH0100307_FAC' ~ -72.140321))
	

#correcting one point that falls outside of LIS watershed
N_load_out$LATITUDE83<-if_else(N_load_out$permit_outfall== 'MA0101214_1', 42.571367,N_load_out$LATITUDE83)
N_load_out$LONGITUDE83<-if_else(N_load_out$permit_outfall== 'MA0101214_1', -72.598121,N_load_out$LONGITUDE83)


    

N_load_out <- N_load_out %>%
  mutate(LATITUDE83 =
           case_when(
             is.na(LATITUDE83) ~ impute_lat,
             !is.na(LATITUDE83) ~LATITUDE83
           )) %>%
   mutate(LONGITUDE83 =
           case_when(
             is.na(LONGITUDE83) ~ impute_lon,
             !is.na(LONGITUDE83) ~LONGITUDE83
           ))
missing_latitude<-N_load_out %>% filter(is.na(LATITUDE83)) %>%
  group_by(permit_outfall)%>%
  summarize(impute_lat=first(impute_lat))
missing_latitude #now no more points are missing
  


```

Loading in NPDES database with NPDES ID to match locations and other data.


```{r}
#HUC8 outfall join
HUC8_outfall_locations <-
  read_csv(file = here('data', 'huc8_outfall_join.csv'))

#select columns
HUC8_outfall_locations <- HUC8_outfall_locations %>%
  select('EXTERNAL_PERMIT_NMBR',
         'PERM_FEATURE_NMBR',
         'huc8',
         'name')

names(HUC8_outfall_locations)

# #need to remove leading zeroes from outfall column
# HUC8_outfall_locations$PERM_FEATURE_NMBR <- sub("^0+", "", HUC8_outfall_locations$PERM_FEATURE_NMBR)
# HUC8_outfall_locations$PERM_FEATURE_NMBR
# 

#create key variables with permit and locations for join
HUC8_outfall_locations$permit_outfall <-
  paste0(HUC8_outfall_locations$EXTERNAL_PERMIT_NMBR,
         "_",
         HUC8_outfall_locations$PERM_FEATURE_NMBR)


#join locations based on permit number and outfall number
dat_join<- left_join(N_load_out, HUC8_outfall_locations, by = "permit_outfall")

dim(N_load_out) #3891
dim(dat_join) #3891
length(unique(as.character(dat_join$permit_outfall))) #51 outfalls with locations
length(unique(as.character(dat_join$key))) #3891 keys

#the number of keys and outfalls with locations are the same as in the last step

#check for NA's in huc8s
missing_huc8<-dat_join %>% filter(is.na(huc8)) %>%
  group_by(permit_outfall)%>%
  summarize(huc8=first(huc8),
            LATITUDE83=first(LATITUDE83),
            LONGITUDE83=first(LONGITUDE83))
missing_huc8


impute_hucs<-data.frame(
  permit_outfall = c(
    'NY0026131_1',
    'NY0026204_1',
    'NY0026701_1',
    'NY0027073_1',
    'NY0206644_2',
    'NY0021750_2',
    'NH0000744_12',
    'NH0000744_13',
    'NH0000744_14',
    'NH0000744_15',
    'NH0000744_16',
    'NH0000744_17',
    'NH0000744_18',
    'NH0000744_19',
    'NH0000744_20',
    'NH0000744_23',
    'NH0000744_24',
    'MA0100889_FAC',
    'MA0101214_1',
    'MA0101281_1',
    'MA0101613_1',
    'MA0101681_2',
    'MA0101818_2',
    'MA0110264_1',
    'NH0000744_22'
    ),
  impute_name = c(
    'East River',
    'East River',
    'Long Island Sound',
    'East River',
    'Long Island Sound',
    'Long Island Sound',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Ammonoosuc River-Connecticut River',
    'Chicopee River',
    'Deerfield River',
    'Deerfield River',
    'Outlet Connecticut River',
    'Housatonic',
    'Ashuelot River-Connecticut River',
    'Ashuelot River-Connecticut River',
    'Ammonoosuc River-Connecticut River'
    )) 

dat_join<- left_join(dat_join, impute_hucs, by = 'permit_outfall')

dat_join <- dat_join %>%
  mutate(name =
           case_when(
             is.na(name) ~ impute_name,
             !is.na(name) ~name
           )) %>%
   mutate(name =
           case_when(
             is.na(name) ~ impute_name,
             !is.na(name) ~name
           ))

missing_huc8<-dat_join %>% filter(is.na(name)) %>%
  group_by(permit_outfall)%>%
  summarize(name=first(name),
            LATITUDE83=first(LATITUDE83),
            LONGITUDE83=first(LONGITUDE83))
missing_huc8 #none are missing now


#clearning up a few last columns
dat_join<- dat_join %>%
  select(-quant_avg,-conc_avg,-days_per_month)


write.csv(dat_join,
          file = here("data", "PCS_data_clean.csv"),
          row.names = F)


```
