==> R/3_ECHO_data_formatting.Rmd <==
---
title: "format ECHO data"
author: "Alex Huddell"
date: "7/28/2021"
output:
  word_document: default
  html_document: default
  pdf_document: default
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#Load libraries and data

```{r }
library(tidyverse)
library(lubridate)
library(here)
library(sf)

col_specifications<-('ccffffffffffddff') #compact string to specify column types

dat<-read_csv(file=unzip(zipfile=here("data","Echo_LIS_complete_data.zip")), col_types = col_specifications)

dat<-distinct(dat) #remove a few duplicate rows

#read in CT data from N exchange program
CT<-read_csv(file=here("data","CT_NX_data.csv"))
CT<-distinct(CT) #remove a few duplicate rows
CT_permits<-unique(CT$permit)

```
#Data exploration visualization
```{r}
#analyzing distributions of Parameter codes
levels(dat$Parameter)

levels<-dat %>%
  group_by(Parameter) %>%
  tally()

ggplot(levels, aes(Parameter,n))+
  geom_col() +
  ylab("count of observations")+
  theme(axis.text.x=element_text(angle = 60, hjust=1))
  
#analyzing distributions of stats codes
levels2<-dat %>%
  group_by(`Statistical Base Long Desc`) %>%
  tally()

levels2 %>% filter(n>1) %>%  ggplot(aes(`Statistical Base Long Desc`,n))+
  geom_col() +
  ylab("count of observations")+
  theme(axis.text.x=element_text(angle = 60, hjust=1))

#looking at distribution of designators
levels(dat$Designator)

levels<-dat %>%
  group_by(Designator) %>%
  tally()

ggplot(levels, aes(Designator,n))+
  geom_col() +
  ylab("count of observations")


#looking at distribution of monitoring locations
levels(dat$`Monitoring Location`)

levels<-dat %>%
  group_by(`Monitoring Location`) %>%
  tally()

ggplot(levels, aes(`Monitoring Location`,n))+
  geom_col() +
  ylab("count of observations")

#looking at distribution of 'column'
levels(dat$Column)

levels<-dat %>%
  group_by(Column) %>%
  tally()

ggplot(levels, aes(Column,n))+
  geom_col() +
  ylab("count of observations")


```


#Standardize data

```{r pressure, echo=FALSE}
#impute zeros for  "No Discharge Code - C" 
dat$`Reported Value` <-
  ifelse(dat$`Reported Value` == "No Discharge Code - C" ,
         0,
         dat$`Reported Value`)#reformating several columns
dim(dat)
names(dat)

#need to remove leading zeroes from outfall column
dat$Outfall <- sub("^0+", "", dat$Outfall)
dat$Outfall


#renaming columns and 
dat<-dat%>%
  rename(date=`Monitoring Period`,
        value=`Reported Value`,
         facility=`Facility Name`,
         permit=`Permit ID`) %>%

# reformatting date and add a key and number of days in month column, then filter out some problematic units
  mutate(date=mdy(date)) %>%
  mutate(key = paste0(permit, "_", Outfall, "_", date)) %>%
  mutate(days_per_month = days_in_month(date))  %>%
  mutate(permit_outfall = as.factor(paste0(permit, "_", Outfall))) %>%
  mutate(permit_outfall_designator=as.factor(paste0(permit,"_",Outfall,"_",Designator))) %>%
  mutate(state=substr(permit,1,2)) %>% #add column with state abbreviation 
  mutate(permit_freq = paste0(permit, "_", `DMR Frequency of Analysis Code1`))


##read outfall locations data, add key, and join to dat
outfall_locations <-
  read_csv(file = here("data", "npdes_outfalls_layer.csv"))

outfall_locations$permit <-outfall_locations$EXTERNAL_PERMIT_NMBR

join<-left_join(dat,outfall_locations, by='permit')

SIC_permit<-join %>%
  group_by(permit) %>%
  summarize(SIC=first(SIC_DESCRIPTIONS),
            SIC_CODES=first(SIC_CODES)) 

SIC_permit_count<- SIC_permit %>%
  group_by(SIC,SIC_CODES) %>%
  tally()%>%
  arrange(desc(n))  

SIC_permit_count%>%
  filter(n>5)%>%
  ggplot(aes(SIC ,n))+
  geom_col() +
  ylab("count of observations")+
  theme(axis.text.x=element_text(angle = 60, hjust=1))

#list of SIC codes to keep?
#	Sewerage Systems--4952
# Colleges And Universities; Elementary And Secondary Schools--8221
#Services; Sewerage Systems--8999 
SIC_CODES_keep<-c('4952','8221','8999', '4952 8221',
                  '9512 4952', '8999 4952')

permits_to_keep<-join %>%
  filter(SIC_CODES %in% SIC_CODES_keep) %>%
  group_by(permit)%>%
  summarize(permit=first(permit),
            facility=first(facility),
            SIC_CODES=first(SIC_CODES))


unique(dat$state)
length(unique(as.character(dat$permit_outfall))) #1270 permit/outfalls
length(levels(as.factor(dat$key))) #159828 keys


```

```{r filter functions}
#some of the flow data for one parameter at many WWTPs is miscoded in ECHO as G, so we separated it and rejoined to the rest of the data (since we do not want to include monitoring location G in other parameters)
flow_ML_G<- dat %>%
   filter(permit %in% (permits_to_keep$permit)) %>%  
   filter(`Monitoring Location`=="G" & Parameter=="50050 - Flow, in conduit or thru treatment plant") %>%
    filter(
    `Statistical Base Long Desc` %in% c(
      "Monthly Avg.",
      "Average Value",
      "Daily Average",
      "30 Day Arithmetic Mean",
      "Monthly Average",
      "30 Day Average", 
      "Average",
      "30 Day Arithmetic"
    )
  ) %>% #by this line, there are  57,703 keys dropped and 535 permit/outfalls dropped 
  filter(Column %in% c("Quantity Average", "Concentration Average")) 

flow_ML_G<-flow_ML_G %>%
  mutate(key2=paste0(permit,date,Designator))  %>%
  filter(! key2=='NY00233112004-03-31M')  %>% #filter out only duplicate
  select(-key2)


dat <- dat %>%
   filter(permit %in% (permits_to_keep$permit)) %>%  
  filter(!permit %in% c('CT0000434','CT0026476','NY0065340','NY0066028')) %>% #removing some industrial permits that seem unrelated to sewerage systems, but some small private WWTPs are still included (like hospitals)
  #filtering to only gross effluent monitoring locations
  filter(`Monitoring Location` %in% c("1", "2", "EA", "EG",  "Y")) %>% #this drops 5,945 keys and 95 permit/outfalls
  #removing infrequently used statistical base codes
  filter(
    `Statistical Base Long Desc` %in% c(
      "Monthly Avg.",
      "Average Value",
      "Daily Average",
      "30 Day Arithmetic Mean",
      "Monthly Average",
      "30 Day Average", 
      "Average",
      "30 Day Arithmetic"
    )
  ) %>% #by this line, there are  57,703 keys dropped and 535 permit/outfalls dropped 
  filter(Column %in% c("Quantity Average", "Concentration Average")) %>%
  filter(!Parameter == "50047 - Flow, maximum during 24 hr period") %>% #by this line, there are 58,062 keys dropped and 538 permit/outfalls dropped
  filter(!permit %in% CT_permits) %>% #this line removes the CT waste water treatment plants because we will use CTDEEP data for those
  #by this line, there are 77,458 keys dropped and 613 permit/outfalls dropped
  # 
  # #removing specific designators that cause duplicates--always removing the less frequent
  # #code when that is clear
  filter(!permit_outfall_designator == 'MA0101508_10_Q') %>% #this one was quarterly vs. subweekly
  #N seems to be the correct designator for the NYC DEP plants 
  filter(!permit_outfall_designator == 'NY0022128_1_M')  %>% #this one I'm not sure, but N was more common
  filter(!permit_outfall_designator == 'NY0026131_1_M')  %>% #this one I'm not sure, but N was more common 
    filter(!permit_outfall_designator == 'NY0026204_1_M')  %>% #this one I'm not sure, but N was more common 
  filter(!permit_outfall_designator == 'NY0026239_1_M')  %>% #this one I'm not sure, but N was more common 
    filter(!permit_outfall_designator == 'NY0026158_1_M')  %>% #this one I'm not sure, but N was more common 	
  filter(!permit_outfall_designator == 'NH0101052_1_B')  %>% #this one was quarterly vs. subweekly 
  drop_na((value)) #drop rows with no value reported

dat<-rbind(dat,flow_ML_G) #add back in flow code exceptions


dat$Parameter <- droplevels(dat)$Parameter # dropping unused levels 
dat$Units <- droplevels(dat)$Units # dropping unused levels 
dat$`Statistical Base Long Desc` <- droplevels(dat)$`Statistical Base Long Desc` #dropping unused levels 
dat$Designator <- droplevels(dat)$Designator # dropping unused levels 

length(unique(as.character(dat$permit_outfall))) #139 permit/outfalls
length(levels(as.factor(dat$key))) #24,045 keys


```

In the next chunk, we reduce the number of units to three--one for flow, one for 
concentration, and another for loading rate. 

```{r}
#standardize units to kg per month, million L per month, or milligram per liter
dat <- dat %>% mutate(
  kg_per_month =
    case_when(
      Units == "(26) - Pounds per Day" ~ value / 2.205 * days_per_month,
      Units == "(01) - Kilograms per Day" ~ value * days_per_month, # converts per day to per month
      Units == "(35) - Grams per Day" ~ value / 1000 * days_per_month, # converts g to kg and per day to per month
      Units == "(2R) - Pounds per Hour" ~ value / 2.205 * 24 * days_per_month, # converts pounds to kg and hours to months
      Units == "(9T) - Million Pounds per Day" ~ value * 10^6 / 2.205 * days_per_month, # converts millions of pounds to kg per month
    )
)

#standardize units to millions of L per month
dat <- dat %>% mutate(
  million_L_per_month =
    case_when(
      Units == "(07) - Gallons per Day" ~ value / 10 ^ 6 * 3.785 * days_per_month,
      #converts from Gallons per day to millions L per month
      Units == "(08) - Cubic Feet per Second" ~ value * 28.317 / 10 ^ 6 * 60 * 60 * 24 * days_per_month,
      # converts cubic feet per second to million L per month
      Units == "(80) - Million Gallons per Month" ~ value * 3.785,
      # converts gallons to L
      Units == "(1X) - Gallons per Hour" ~ value / 10 ^ 6 * 3.785 * 24 * days_per_month,
      # converts gallons per hour to millions of L per month
      Units == "(8D) - Gallons per Month" ~ value / 10 ^ 6 * 3.785,
      # converts gallons to millions of L
      Units == "(03) - Million Gallons per Day" ~ value * 3.785 * days_per_month,
      # converts gallons to L and day to month
      Units == "(78) - Gallons per Minute" ~ value / 10 ^ 6 * 3.785 * 60 * 24 * days_per_month,
      # converts gallons to millions of L and minute to month
    )
)

#converting all N parameters to mass terms of nitrogen and not others (i.e. ammonia, nitrite, etc.)


dat$value <-  ifelse(
  dat$Parameter == "00608 - Nitrogen, ammonia dissolved",
  dat$value * 14.0067 / 17.031,
  dat$value
)

dat$value <- ifelse(
  dat$Parameter == "34726 - Nitrogen, ammonia, total (as NH3)",
  dat$value * 14.0067 / 17.031,
  dat$value
)



#standardize concentrations to milligrams N per liter  
dat <- dat %>% mutate(
  mg_N_per_L=
    case_when(
   Units == "(19) - Milligrams per Liter" ~ value,
   Units == "(28) - Micrograms per Liter" ~ value / 1000, #convert from micrograms to milligram
    )
  )


```
Next, we reduce the number of nitrogen to the following: total nitrogen "TN", organic N "ON",
total Kjeldahl nitrogen "TKN", ammonia "NH3", nitrite "NO2", nitrate "NO3", and NO2NO3 for nitrate and nitrite combined.
We reduce the flow parameters to the following: flow rate 'flow_rate", 
maximum flow rate "flow_max_24h", total flow "flow_total", and dry flow "flow_dry".

```{r}
#cleaning up duplicate nitrogen and flow parameters 
dat$param_simple <- recode(dat$Parameter,
    #Nitrogen parameters
    "00600 - Nitrogen, total (as N)" = "TN",
    "00605 - Nitrogen, organic total (as N)" = "ON",
    "00610 - Nitrogen, ammonia total (as N)" = "NH3",
    "00615 - Nitrogen, nitrite total (as N)" = "NO2",
    "00620 - Nitrogen, nitrate total (as N)" = "NO3",
    "00625 - Nitrogen, Kjeldahl, total (as N)"= "TKN",
    "51445 - Nitrogen, Total" = "TN",
    "00608 - Nitrogen, ammonia dissolved" = "NH3",
    "00630 - Nitrite + Nitrate total (as N)" = "NO3NO2",
    "00630 - Nitrite Plus Nitrate Total 1 Det. (As N)"= "NO3NO2",
    "00609 - Ammonia nitrogen, total, (as N) 30 day" = "NH3",
    "49579 - Nitrogen, total Kjeldahl" = "TKN",
    "51447 - Nitrogen, Nitrite  Total" = "NO2",
    "51448 - Nitrogen, Nitrate  Total" = "NO3",
    "51449 - Nitrogen, Kjeldahl  Total" = "TKN",
    "51446 - Nitrogen, Ammonia  Total" = "NH3",
    "51087 - Nitrogen, Kjeldahl, total (TKN) (water)" = "TKN",
    "51489 - Nitrogen, Total as NO3 + NH3" = "NO3NH3",
    "34726 - Nitrogen, ammonia, total (as NH3)" = "NH3",
    "51085 - Nitrogen, ammonia (NH3-N)" = "NH3",
    "51425 - Nitrogen, Total As N" = "TN",
    "51450 - Nitrite Plus Nitrate Total" = "NO3NO2", 
    #flow parameters
    "00056 - Flow rate" = "flow_rate",
    "74076 - Flow" = "flow_rate",
    "00058 - Flow rate" = "flow_rate",
    "82220 - Flow, total" = "flow_total",
    "50050 - Flow, in conduit or thru treatment plant" ="flow_rate",
    "51061 - Flow (dry weather)" = "flow_total"
  )

#analyzing distribution of parameters
levels(dat$param_simple)

levels<-dat %>%
  group_by(param_simple) %>%
  tally()

ggplot(levels, aes(param_simple,n))+
  geom_col() +
  ylab("count of observations")


levels(dat$Parameter)

levels<-dat %>%
  group_by(Parameter) %>%
  tally()

ggplot(levels, aes(Parameter,n))+
  geom_col() +
  ylab("count of observations")+
  theme(axis.text.x=element_text(angle = 60, hjust=1))


```

#Standardizing statistics codes
For now,I excluded ALL but four statistical base codes because the others relate to maxima or other irrelevant data.
In the future, we may want to add back in some of the other statistics.

```{r}

dat$stat <- recode(dat$`Statistical Base Long Desc`,
    "Monthly Avg." = "mean",
    "Average Value" = "mean",
    "Daily Average" = "mean", 
    "30 Day Arithmetic Mean" = "mean",
     "Monthly Average" = "mean",
      "30 Day Average" = "mean",
      "Average" = "mean",
      "30 Day Arithmetic" = "mean")

dat$stat <- droplevels(dat)$stat # dropping unused levels from dat$stat

levels(dat$stat)



```

Tallying up duplicate values that complicate reshaping the data from long to wide formatting. 

```{r, echo=T}
#look at how many rows there are for each key
n_row<-dat %>%
  group_by(key)%>%
  tally() %>%
  arrange(desc(n))

hist(n_row$n) #most observations have 1-2 rows only, but a few have up to 12 rows

```
First, we'll separate the keys for which kg N/month loads already exist from those which don't, check for duplicates, then pivot data from long to wide format.

```{r}
#subset of data that report loads
N_load<- dat %>%
  group_by(key) %>%
  filter(!is.na(kg_per_month)) %>% 
  #removing unwanted columns
  select(
    !c(value,
       Parameter,
      `Received Date`,
      `Monitoring Location`,
      Units,
      `Statistical Base Long Desc`,
      Column,
      days_per_month,
      Designator,
      `Seasonal Indicator`,
      Limit,
      permit_freq,
      `DMR Frequency of Analysis Code1`,
      `Limit Sample Type`)
  ) %>%
  distinct() #removing any duplicate rows


#check if there are duplicates per key and one parameter
dup<-N_load %>%
  group_by(key,param_simple)%>%
  #filter(param_simple=="NO3NH3") %>%
  tally() %>%
  arrange(desc(n))
dup

#there are no more duplicates here, but there are multiple parameters per key 

#pivoting the TN and NH3 observations into wide format and removing flow and concentration columns
N_load_a<-N_load %>%
   pivot_wider(names_from = param_simple,
              values_from = kg_per_month) %>%
  select(-million_L_per_month,- mg_N_per_L,-stat , -permit_outfall_designator) %>%
  rename(kg_N_TN_per_month=TN,
         kg_N_NH3_per_month=NH3,
         kg_N_TKN_per_month=TKN,
         kg_N_NO3NO2_per_month=NO3NO2,
         kg_N_NO3NH3_per_month=NO3NH3,
         kg_N_NO3_per_month=NO3,
         kg_N_NO2_per_month=NO2)

names(N_load_a)

dup<-N_load_a %>%
  group_by(key)%>%
  tally() %>%
  filter(n>2) %>%
  arrange(desc(n))
dup #there are no duplicate

length(unique(as.character(N_load$key)))
length(unique(as.character(N_load_a$key))) 
#there are the same number of rows before and after the pivot


length(unique(as.character(N_load_a$permit_outfall))) #68 outfalls included here


```

Here we assess and then resolve duplicates by selecting only 1 flow parameter for each outfall and date, pivot the data wider calculate loads based on the concentrations and flow data, and remove the concentration and flow columns after we use them.


```{r}
#function to id columns with all NAs
not_all_na <- function(x) any(!is.na(x))

#review duplicates
keys_with_N<- dat %>%
  group_by(key) %>%
  filter(!any(!is.na(kg_per_month)))%>% #this removes any keys which do report in N loads
  filter(any(Parameter %in% c(
      "00600 - Nitrogen, total (as N)",
      "00605 - Nitrogen, organic total (as N)",
      "00610 - Nitrogen, ammonia total (as N)",
      "00615 - Nitrogen, nitrite total (as N)",
      "00620 - Nitrogen, nitrate total (as N)",
      "51445 - Nitrogen, Total",
      "00608 - Nitrogen, ammonia dissolved",
      "00630 - Nitrite + Nitrate total (as N)",
      "00609 - Ammonia nitrogen, total, (as N) 30 day",
      "49579 - Nitrogen, total Kjeldahl",
      "51447 - Nitrogen, Nitrite  Total",
      "51448 - Nitrogen, Nitrate  Total" ,
      "51449 - Nitrogen, Kjeldahl  Total" ,
      "51446 - Nitrogen, Ammonia  Total",
      "51087 - Nitrogen, Kjeldahl, total (TKN) (water)",
      "51489 - Nitrogen, Total as NO3 + NH3" ,
      "34726 - Nitrogen, ammonia, total (as NH3)",
      "51085 - Nitrogen, ammonia (NH3-N)",
      "51425 - Nitrogen, Total As N",
      "51450 - Nitrite Plus Nitrate Total"
    )) #this removes any keys which do report any other N parameters
  ) 
keys_with_N


dup<-keys_with_N %>%
  group_by(key) %>%
  filter(any(is.na(kg_per_month))) %>%
  group_by(key, param_simple)%>%
  tally() %>%
  arrange(desc(n))
dup #there are no more duplicates with N parameters

#separate out concentration data so that we can drop keys empty concentrations
N_params<- keys_with_N %>%
group_by(key) %>%
  filter(!any(!is.na(kg_per_month)))%>% #this removes any keys which do report in N loads
  filter(!Parameter == '50050 - Flow, In Conduit Or Thru Treatment Plant') %>%
  filter(!Parameter == '00056 - Flow rate') %>%
  filter(!Parameter == '74076 - Flow') %>%
  filter(!Parameter == '00058 - Flow rate') %>%
  filter(!Parameter == '51061 - Flow (dry weather)') %>%
  drop_na(mg_N_per_L)%>%
distinct()

#separate out flows
flow_all<- keys_with_N %>%
group_by(key) %>%
  filter(!any(!is.na(kg_per_month)))%>% #this removes any keys which do report in N loads
  filter(Parameter %in% 
           c('50050 - Flow, in conduit or thru treatment plant', 
             '00056 - Flow rate', 
              '74076 - Flow',
              '00058 - Flow rate',
             '51061 - Flow (dry weather)') 
         )

#bind all observations back together
flow_cleaned<-rbind(flow_all,N_params)
 
flow_cleaned <- flow_cleaned %>%
  #removing unwanted columns
  select(
    !c(value,
       Parameter,
      `Received Date`,
      `Monitoring Location`,
      Units,
      `Statistical Base Long Desc`,
      Column,
      days_per_month,
      Designator,
      `Seasonal Indicator`,
      Limit,
      `DMR Frequency of Analysis Code1`,
      `Limit Sample Type`)
  ) %>%
  distinct() 

 
 #review duplicates
dups2<-flow_cleaned %>%
  group_by(key, param_simple)%>%
  tally() %>%
  arrange(desc(n))
dups2 #there are no more duplicates


```

Calculate loads for keys which do not report loads.

```{r}

#subset of data that DO NOT report loads, then
#pivot wider and then remove columns with only NAs (since some parameter/unit combinations that don't make sense)
flow_N_concentration <- flow_cleaned %>%
  select(-permit_freq,-permit_outfall_designator) %>%
  group_by(key) %>%
  filter(any(is.na(kg_per_month))) %>%
  pivot_wider(names_from = param_simple,
              values_from = c(mg_N_per_L, million_L_per_month)) %>%
  select_if(not_all_na) 

names(flow_N_concentration)

#review duplicates
dups3<-flow_N_concentration %>%
  group_by(key)%>%
  tally() %>%
  arrange(desc(n))
dups3 #there are no more duplicates

#calculating N loads
N_load_b <- flow_N_concentration %>%
  mutate(kg_N_TN_per_month = mg_N_per_L_TN * million_L_per_month_flow_rate) %>%
  mutate(kg_N_NH3_per_month = mg_N_per_L_NH3 * million_L_per_month_flow_rate) %>%
  mutate(kg_N_NO2_per_month = mg_N_per_L_NO2 * million_L_per_month_flow_rate) %>%
  mutate(kg_N_NO3_per_month = mg_N_per_L_NO3 * million_L_per_month_flow_rate) %>%
  mutate(kg_N_TKN_per_month = mg_N_per_L_TKN * million_L_per_month_flow_rate) %>%
  #mutate(kg_N_NO3NO2_per_month = mg_N_per_L_NO3NO2 * million_L_per_month_flow_rate) %>% #this is no longer in the data
select(
    -mg_N_per_L_TN,
    -mg_N_per_L_NH3,
    -mg_N_per_L_NO2,
    -mg_N_per_L_NO3,
  #  -mg_N_per_L_NO3NO2,
    -mg_N_per_L_TKN,
    -million_L_per_month_flow_rate,
    -stat
  ) 

#check if any rows contain both NO3 or NO2 and NO3NO2 before we combine them
#filter(N_load_b, !is.na(kg_N_NO3_per_month) & !is.na(kg_N_NO3NO2_per_month))
#filter(N_load_b, !is.na(kg_N_NO2_per_month) & !is.na(kg_N_NO3NO2_per_month))
#there are none, so we can add them without double counting


#sum up other forms of N apart from TN
N_load_b <- N_load_b %>%
  mutate(kg_N_sum_per_month =
           case_when(
             all(
               is.na(kg_N_NH3_per_month),
               is.na(kg_N_NO3_per_month),
               is.na(kg_N_NO2_per_month)#,
              # is.na(kg_N_NO3NO2_per_month)
               ) ~ NA_real_,
             is.na(kg_N_TKN_per_month) ~ sum(
               kg_N_NH3_per_month,
               kg_N_NO3_per_month,
               kg_N_NO2_per_month,
            #   kg_N_NO3NO2_per_month,
               na.rm = TRUE
             ),
             !is.na(kg_N_TKN_per_month) ~ sum(
               kg_N_TKN_per_month,
               kg_N_NO3_per_month,
               kg_N_NO2_per_month,
              # kg_N_NO3NO2_per_month,
               na.rm = TRUE
             )
           )) 

#remove intermediate columns we no longer need
N_load_b <- N_load_b %>%
  select(
    -kg_N_NH3_per_month,
    -kg_N_NO3_per_month,
    -kg_N_NO2_per_month,
  #  -kg_N_NO3NO2_per_month,
    -kg_N_TKN_per_month
  )

#combine the different N columns; use total N when reported, or the sum of all other
#forms of N if TN is not reported

N_load_b$kg_N_TN_per_month <-  ifelse(!is.na(N_load_b$kg_N_TN_per_month),
    N_load_b$kg_N_TN_per_month,
    N_load_b$kg_N_sum_per_month )

#remove other unnecessary columns
N_load_b<- N_load_b %>%
  select(-kg_N_sum_per_month)

names(N_load_b)
length(unique(as.character(N_load_b$permit_outfall))) #22 outfalls included here

#for N_load_a
#combine the different N columns; use total N when reported, or the sum of all other
#forms of N if TN is not reported

#check if any rows contain both NO3 or NO2 and NO3NO2 before we combine them
filter(N_load_a, !is.na(kg_N_NO3_per_month) & !is.na(kg_N_NO3NO2_per_month))
#there are none, so we can add them without double counting
filter(N_load_a, !is.na(kg_N_NO2_per_month) & !is.na(kg_N_NO3NO2_per_month)) #26 rows have both, so we have to subtract NO2 from kg_N_NO3NO2_per_month for those rows

#checking for both NH3 or NO3 and NH3NO3
filter(N_load_a, !is.na(kg_N_NO3_per_month) & !is.na(kg_N_NO3NH3_per_month)) 
#there are none, so we can add them without double counting
filter(N_load_a, !is.na(kg_N_NH3_per_month) & !is.na(kg_N_NO3NH3_per_month))
#there are none, so we can add them without double counting

#checking for both TKN  or NO3 and NH3NO3
filter(N_load_a, !is.na(kg_N_TKN_per_month) & !is.na(kg_N_NO3NH3_per_month))
#there are none, so we can add them without double counting

#isolating rows that report both NO2 and NO3NO2 to remove NO2
N_load_a_1<-N_load_a %>%
  filter(!is.na(kg_N_NO2_per_month) & !is.na(kg_N_NO3NO2_per_month))
  
N_load_a_1$kg_N_NO2_per_month<-0 #imputing zeroes into the NO2 column to avoid double counting
    
keys_N_load_a_1<-(N_load_a_1$key)

N_load_a_2<-N_load_a %>%
  filter(!key %in% keys_N_load_a_1) #selecting all other rows to rejoin

N_load_a<-rbind(N_load_a_1, N_load_a_2)

#calculate sum of other forms of N
N_load_a$kg_N_sum_per_month <-
  ifelse(
    !is.na(N_load_a$kg_N_TKN_per_month),
    (
      N_load_a$kg_N_TKN_per_month + N_load_a$kg_N_NO3_per_month +
        N_load_a$kg_N_NO2_per_month + N_load_a$kg_N_NO3NO2_per_month +
        N_load_a$kg_N_NO3NH3_per_month
    ),
    (
      N_load_a$kg_N_NH3_per_month + N_load_a$kg_N_NO3_per_month +
        N_load_a$kg_N_NO2_per_month +  N_load_a$kg_N_NO3NO2_per_month +
        N_load_a$kg_N_NO3NH3_per_month
    )
  )


#remove intermediate columns we no longer need
N_load_a <- N_load_a %>%
  select(
    -kg_N_NH3_per_month,
    -kg_N_NO3_per_month,
    -kg_N_NO2_per_month,
    -kg_N_NO3NO2_per_month,
    -kg_N_TKN_per_month,
    -kg_N_NO3NH3_per_month
  )

#combine the different N columns; use total N when reported, or the sum of all other
#forms of N if TN is not reported

N_load_a$kg_N_TN_per_month <-  ifelse(!is.na(N_load_a$kg_N_TN_per_month),
    N_load_a$kg_N_TN_per_month,
    N_load_a$kg_N_sum_per_month )

#remove other unnecessary columns
N_load_a<- N_load_a %>%
  select(-kg_N_sum_per_month)
names(N_load_a)

#combine dataframes
N_load<-rbind(N_load_a,N_load_b)

N_load<-drop_na(N_load,kg_N_TN_per_month)

length(unique(as.character(N_load$permit_outfall))) #70 outfalls with/without locations
nrow(N_load) #8330

summary(N_load$kg_N_TN_per_month)

```

#Format and add in the CT data

```{r}

CT<-read_csv(file=here("data","CT_NX_data.csv"))
CT<-distinct(CT) #remove a few duplicate rows

#formatting CT data to join with rest of ECHO data
CT_summary<-CT %>%
  mutate(date=mdy(date)) %>%
  mutate(month_year=format(as.Date(date), "%Y-%m")) %>%
  mutate(days_date = day(date),
         days_per_month = days_in_month(date)) %>%
  mutate(Outfall=rep(1)) %>%
 # mutate(TN_kg_d=(if_else(TN_kg_d==3792403.8,1400/2.2,TN_kg_d))) %>% #imputing values for some obvious typos/errors in data
  group_by(permit,Outfall,month_year) %>%
  arrange(date)%>%
  mutate(date_sequence = row_number())%>%
  mutate(day_interval= case_when( 
    date_sequence == 1 ~ days_date,
    date_sequence > 1 ~ days_date-dplyr::lag(days_date, n=1)
    )) %>% #computing the number of days in each reporting date interval
  mutate(TN_kg_interval=day_interval*TN_kg_d) %>% #calculating N load per interval
  group_by(permit,Outfall,month_year,days_per_month) %>%
  summarise(kg_N_TN_total=sum(TN_kg_interval),
            day_interval_total=sum(day_interval)) %>%
  mutate(permit_outfall=paste0(permit,"_",Outfall),
         month_rate=kg_N_TN_total/day_interval_total) %>%
  mutate(kg_N_TN_per_month=month_rate*days_per_month) %>%
  mutate(key= paste0(permit, "_", Outfall, "_", month_year)) %>%
  select(-month_rate, -day_interval_total) 
#removing last three months of CT data which are not included in the rest of the ECHO data

names(CT)
permit_facility<-CT %>%
  group_by(permit) %>%
  summarize(permit = first(permit),
            facility = first(facility))
nrow(CT_summary) #17934
CT_summary<-left_join(CT_summary, permit_facility)
nrow(CT_summary) #17934

CT_summary<-CT_summary %>% select(-kg_N_TN_total,-days_per_month)

N_load<- N_load %>%
  mutate(month_year=format(as.Date(date), "%Y-%m")) %>%
  select(-date)
N_load$Outfall<-as.character(N_load$Outfall)

CT_summary$Outfall<-as.character(CT_summary$Outfall)
CT_summary$state<-rep('CT')

names(N_load)
names(CT_summary)

#join locations based on permit number and outfall number
N_load_CT_join <- rbind(N_load, CT_summary)
N_load_CT_join<-distinct(N_load_CT_join)

dim(N_load_CT_join) #26264
dim(N_load) #8330    
dim(CT_summary) #17934     
length(unique(as.character(N_load_CT_join$permit_outfall))) #151 outfalls with locations
length(unique(as.character(N_load_CT_join$key))) #26264 keys


```

Loading in NPDES database with NPDES ID to match locations and other data.

```{r outfall location join}
outfall_locations <-
  read_csv(file = here("data", "npdes_outfalls_layer.csv"))
names(outfall_locations)
outfall_locations<-outfall_locations %>%
  select(EXTERNAL_PERMIT_NMBR,
         FACILITY_NAME,
         SIC_CODES,
         SIC_DESCRIPTIONS,
         PERM_FEATURE_NMBR,
         STATE_WATER_BODY_NAME,
         LATITUDE83,
         LONGITUDE83)
names(N_load_CT_join)
#need to remove leading zeroes from outfall column
N_load_CT_join$Outfall <- sub("^0+", "", N_load_CT_join$Outfall)
N_load_CT_join$Outfall
#create key variables with permit and locations for join
N_load_CT_join$permit_outfall <- paste0(N_load_CT_join$permit, "_", N_load_CT_join$Outfall)
outfall_locations$permit_outfall <-
  paste0(outfall_locations$EXTERNAL_PERMIT_NMBR,
         "_",
         outfall_locations$PERM_FEATURE_NMBR)
#join locations based on permit number and outfall number
dim(N_load_CT_join) #25392          
length(unique(N_load_CT_join$permit_outfall)) #150 outfalls with/without locations
length(unique(N_load_CT_join$key)) #25392      keys


N_load_out <- left_join(N_load_CT_join, outfall_locations, by = "permit_outfall")
dim(N_load_out) #25392     
length(unique(as.character(N_load_out$permit_outfall))) #150 outfalls with locations
length(unique(as.character(N_load_out$key))) #25392       keys
#the data are the same dimensions before and after the join

#check for NA's in latitude
summary(N_load_out$LATITUDE83) #456 missing values
missing_latitude<-filter(N_load_out, is.na(LATITUDE83))
missing_latitude<-N_load_out %>% filter(is.na(LATITUDE83)) %>%
  group_by(permit_outfall)%>%
  summarize(LATITUDE83=first(LATITUDE83))
missing_latitude #5 outfalls missing coordinates

#impute missing values based on other outfalls from same permit
N_load_out<-N_load_out %>% 
  mutate(impute_lat= case_when(
    permit_outfall=='MA0101508_7' ~ 42.151998,
    permit_outfall=='MA0101630_9' ~ 42.190537,
    permit_outfall== 'NY0021750_2' ~ 40.948583,	
    permit_outfall== 'NY0206644_2' ~ 40.924472,	
    permit_outfall== 'NY0206644_3' ~ 40.951917)) %>%
   mutate(impute_lon= case_when(
    permit_outfall=='MA0101508_7' ~ -72.624415,
    permit_outfall=='MA0101630_9' ~ -72.609792,
    permit_outfall== 'NY0021750_2' ~ -73.075639,
    permit_outfall== 'NY0206644_2' ~ -73.121139,
    permit_outfall== 'NY0206644_3' ~ -73.073139))


N_load_out <- N_load_out %>%
  mutate(LATITUDE83 =
           case_when(
             is.na(LATITUDE83) ~ impute_lat,
             !is.na(LATITUDE83) ~LATITUDE83
           )) %>%
   mutate(LONGITUDE83 =
           case_when(
             is.na(LONGITUDE83) ~ impute_lon,
             !is.na(LONGITUDE83) ~LONGITUDE83
           ))

missing_latitude<-N_load_out %>% filter(is.na(LATITUDE83)) %>%
  group_by(permit_outfall)%>%
  summarize(LATITUDE83=first(LATITUDE83))
missing_latitude #no coordinates are missing now

  

```

Loading in NPDES database with NPDES ID to match locations and other data.


```{r}
#HUC8 outfall join
HUC8_outfall_locations <-
  read_csv(file = here('data', 'huc8_outfall_join.csv'))
#select columns
HUC8_outfall_locations <- HUC8_outfall_locations %>%
  select('EXTERNAL_PERMIT_NMBR',
         'PERM_FEATURE_NMBR',
         'huc8',
         'name')
names(HUC8_outfall_locations)

#need to remove leading zeroes from outfall column
HUC8_outfall_locations$PERM_FEATURE_NMBR <- sub("^0+", "", HUC8_outfall_locations$PERM_FEATURE_NMBR)
HUC8_outfall_locations$PERM_FEATURE_NMBR

#create key variables with permit and locations for join
HUC8_outfall_locations$permit_outfall <-
  paste0(HUC8_outfall_locations$EXTERNAL_PERMIT_NMBR,
         "_",
         HUC8_outfall_locations$PERM_FEATURE_NMBR)
HUC8_outfall_locations$permit_outfall
N_load_out$permit_outfall

#join locations based on permit number and outfall number
dat_join<- left_join(N_load_out, HUC8_outfall_locations, by = "permit_outfall")

dim(N_load_out) #25392        
dim(dat_join) #25392     #same length as before join

length(unique(as.character(dat_join$permit_outfall))) #150 outfalls with locations
length(unique(as.character(dat_join$key))) #25392     keys
#the number of keys and outfalls with locations are the same as in the last step

#check for NA's in huc8s
missing_huc8<-dat_join %>% filter(is.na(huc8)) %>%
  group_by(permit_outfall)%>%
  summarize(name=first(name),
            LATITUDE83=first(LATITUDE83),
            LONGITUDE83=first(LONGITUDE83))
missing_huc8 

impute_hucs<-data.frame(
  permit_outfall = c(
    'NY0026131_1',
    'NY0026204_1',
    'NY0026701_1',
    'NY0027073_1',
    'NY0206644_2',
    'NY0206644_3',
    'NY0021750_2',
    'MA0101508_7',
    'MA0101630_9',
    'NH0000744_1'),
  impute_name = c(
    'East River',
    'East River',
    'Long Island Sound',
    'East River',
    'Long Island Sound',
    'Long Island Sound',
    'Long Island Sound',
    'Chicopee River', 
    'Ashuelot River-Connecticut River',
    'Ammonoosuc River-Connecticut River')) 

dat_join<- left_join(dat_join, impute_hucs, by = 'permit_outfall')

dat_join <- dat_join %>%
  mutate(name =
           case_when(
             is.na(name) ~ impute_name,
             !is.na(name) ~name
           )) %>%
   mutate(name =
           case_when(
             is.na(name) ~ impute_name,
             !is.na(name) ~name
           ))
#check for NA's in huc8s
sum(is.na(dat_join$name))

#checking duplicates
dim(dat_join)
dat_join<-distinct(dat_join)
dim(dat_join)

write_csv(dat_join,
          file = here("data", "ECHO_data_clean.csv"))

```
